requirements.txt
fastapi
uvicorn
openai
python-dotenv
langchain
langchain-community
chromadb
langchain-openai
